courier/requirements.txt
python-dotenv
redis
bcrypt